Refactor reservation serializers to own sub package
It was getting bit heavy file.

Refs TILA-2074
