+++ api/graphql/reservations/reservation_serializers/memo_serializers.py
@@ -1,0 +1,12 @@
+from api.graphql.base_serializers import PrimaryKeySerializer
+from reservations.models import Reservation
+
+
+class ReservationWorkingMemoSerializer(PrimaryKeySerializer):
+    class Meta:
+        model = Reservation
+        fields = ["pk", "working_memo"]
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.fields["pk"].help_text = "Primary key of the reservation"
